requirements.txt
Flask>=2.3.3
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.0.5
Flask-Migrate>=4.0.4
Flask-Session>=0.5.0
Flask-WTF>=1.1.1
WTForms>=3.0.1
python-dotenv>=1.0.1
Werkzeug>=2.3.7
Jinja2>=3.1.2
itsdangerous>=2.1.2
click>=8.1.3
Flask-Mail>=0.9.1
mysqlclient>=2.2.0
pymysql>=1.1.0
requests>=2.31.0
google-generativeai>=0.3.2

Flask-Login==0.5.0
SQLAlchemy==1.4.49
mysql-connector-python==8.0.26
PyMySQL==1.0.2
email-validator==1.1.3
Pillow>=10.0.0
Flask-WTF==0.15.1
twilio==8.12.0
pyotp==2.9.0
cryptography>=41.0.0 